requirements.txt
flask~=2.2.2
requests~=2.28.1
pandas~=1.3.5
torch~=1.13.0
Pillow~=9.3.0
Werkzeug~=2.2.2
numpy~=1.21.6
albumentations~=1.3.0